requirements.txt
appdirs==1.4.4
asgiref==3.3.4
distlib==0.3.1
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
filelock==3.0.12
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.4
